backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
email-validator==2.2.0
pytest==8.3.2
httpx==0.27.2

ortools==9.10.4067

